Some fixes and added message sending by ai skill

+++ lib/message_sender.rb
@@ -56,6 +56,9 @@ class MessageSender
     return if delete_bot_message.nil? || bot_messages.empty?
 
     msg_on_destroy = find_msg_on_destroy
+
+    return if @tg_user.id.to_i != msg_on_destroy.chat_id.to_i # Messages deletion only for current tg user
+
     bot.api.delete_message(message_id: msg_on_destroy.message_id, chat_id: msg_on_destroy.chat_id)
   end
